:sparkles: Created new geolocation library, subsequently broke some functionality in sign-up...
