+++ libs/authentication/lib/pages/login_page.dart
@@ -6,7 +6,6 @@ import 'package:flutter/material.dart';
 import 'package:general/theme.dart';
 import 'package:general/widgets/divider.dart';
 import 'package:redux_comp/actions/init_amplify_action.dart';
-import 'package:redux_comp/actions/get_address_action.dart';
 import 'package:redux_comp/redux_comp.dart';
 import '../widgets/button.dart';
 import '../widgets/link.dart';
@@ -252,21 +251,21 @@ class _Factory extends VmFactory<AppState, LoginPage> {
         dispatchLoginAction: (String email, String password) => dispatch(
           LoginAction(email, password),
         ),
-        dispatchGetAddressAction: () => dispatch(
-          GetAddressAction(),
-        ),
+        // dispatchGetAddressAction: () => dispatch(
+        //   GetAddressAction(),
+        // ),
       );
 }
 
 // view model
 class _ViewModel extends Vm {
   final void Function(String, String) dispatchLoginAction;
-  final void Function() dispatchGetAddressAction;
+  // final void Function() dispatchGetAddressAction;
   final VoidCallback pushSignUpPage;
 
   _ViewModel({
     required this.dispatchLoginAction,
-    required this.dispatchGetAddressAction,
+    // required this.dispatchGetAddressAction,
     required this.pushSignUpPage,
   }); // implementinf hashcode
 }
