+++ libs/geolocation/lib/geolocation.dart
@@ -1,0 +1,7 @@
+library geolocation;
+
+/// A Calculator.
+class Calculator {
+  /// Returns [value] plus 1.
+  int addOne(int value) => value + 1;
+}
