+++ libs/geolocation/pubspec.yaml
@@ -1,47 +1,41 @@
-name: redux_comp
-description: redux componenets for the project
+name: geolocation
+description: A service library relating to geolocation functionality (Places API, geocoding)
 version: 0.0.1
-homepage:
 publish_to: none
+homepage:
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.1 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  geolocation:
-    path: ../geolocation
+  http: ^0.13.4
+
+  async_redux: ^15.0.0
+
+  redux_comp: 
+    path: '../redux_comp'
+
 
   geocoding: ^2.0.4
   geolocator: ^9.0.0
 
-  # amplify
   amplify_flutter: ^0.5.1
-  amplify_auth_cognito: ^0.5.1
   amplify_api: ^0.5.1
 
-  # redux
-  async_redux: ^15.0.0
-
-  # other
-  uuid: ^3.0.6
-  jwt_decode: ^0.3.1
-
-  amplify:
-    path: '../amplify'
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  flutter_lints: ^2.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
-# The following section is specific to Flutter.
+# The following section is specific to Flutter packages.
 flutter:
 
   # To add assets to your package, add an assets section, like this:
@@ -53,7 +47,7 @@ flutter:
   # https://flutter.dev/assets-and-images/#from-packages
   #
   # An image asset can refer to one or more resolution-specific "variants", see
-  # https://flutter.dev/assets-and-images/#resolution-aware.
+  # https://flutter.dev/assets-and-images/#resolution-aware
 
   # To add custom fonts to your package, add a fonts section here,
   # in this "flutter" section. Each entry in this list should have a
