+++ libs/redux_comp/lib/actions/get_address_action.dart
@@ -1,140 +1,30 @@
-import 'dart:convert';
-
-import 'package:amplify_api/amplify_api.dart';
-import 'package:amplify_flutter/amplify_flutter.dart';
 
 import '../app_state.dart';
 import 'package:async_redux/async_redux.dart';
-import 'package:http/http.dart' as http;
+
+import '../models/geolocation/suggestion_model.dart';
+import 'package:geolocation/place_api_service.dart';
 
 class GetAddressAction extends ReduxAction<AppState> {
-  // String input;
-  // String sessionToken;
+  String input;
+  PlaceApiService placeApi;
 
-  // GetAddressAction(this.input, this.sessionToken);
+  GetAddressAction(this.input, this.placeApi);
 
   @override
   Future<AppState?> reduce() async {
-    PlaceApiProvider p = PlaceApiProvider(1234);
-
-    List<Suggestion> l = await p.fetchSuggestions("318 The Rand");
-
-    l.toString();
-
-    return null;
-  }
-}
-
-class Place {
-  String? streetNumber;
-  String? street;
-  String? city;
-  String? zipCode;
-
-  Place({
-    this.streetNumber,
-    this.street,
-    this.city,
-    this.zipCode,
-  });
-
-  @override
-  String toString() {
-    return 'Place(streetNumber: $streetNumber, street: $street, city: $city, zipCode: $zipCode)';
-  }
-}
-
-class Suggestion {
-  final String placeId;
-  final String description;
-
-  Suggestion(this.placeId, this.description);
-
-  @override
-  String toString() {
-    return 'Suggestion(description: $description, placeId: $placeId)';
-  }
-}
-
-class PlaceApiProvider {
-  final client = http.Client();
-
-  final sessionToken;
-  Future<String> getApiKey() async {
-    String graphQLDocument = '''query {
-      viewKey
-    }''';
-
-    final request = GraphQLRequest(
-      document: graphQLDocument,
-    );
     try {
-      final data = jsonDecode(
-          (await Amplify.API.mutate(request: request).response).data);
-      return data['viewKey'];
+      List<Suggestion> l = await placeApi.fetchSuggestions(input);
+      return state.replace(
+        geo: state.geo!.replace(
+          suggestions: l
+        )
+      );
     } catch (e) {
-      return "Fetching Key Failed";
+      return null;
     }
-  }
 
-  PlaceApiProvider(this.sessionToken);
 
-  Future<List<Suggestion>> fetchSuggestions(String input) async {
-    final apiKey = await getApiKey();
-    final request =
-        'https://maps.googleapis.com/maps/api/place/autocomplete/json?input=$input&types=address&language=en&components=country:za&key=$apiKey&sessiontoken=$sessionToken';
-    final response = await client.get(Uri.parse(request));
-
-    if (response.statusCode == 200) {
-      final result = json.decode(response.body);
-      if (result['status'] == 'OK') {
-        // compose suggestions in a list
-        return result['predictions']
-            .map<Suggestion>((p) => Suggestion(p['place_id'], p['description']))
-            .toList();
-      }
-      if (result['status'] == 'ZERO_RESULTS') {
-        return [];
-      }
-      throw Exception(result['error_message']);
-    } else {
-      throw Exception('Failed to fetch suggestion');
-    }
-  }
-
-  Future<Place> getPlaceDetailFromId(String placeId) async {
-    final apiKey = await getApiKey();
-    final request =
-        'https://maps.googleapis.com/maps/api/place/details/json?place_id=$placeId&fields=address_component&key=$apiKey&sessiontoken=$sessionToken';
-    final response = await client.get(Uri.parse(request));
-
-    if (response.statusCode == 200) {
-      final result = json.decode(response.body);
-      if (result['status'] == 'OK') {
-        final components =
-            result['result']['address_components'] as List<dynamic>;
-        // build result
-        final place = Place();
-        for (var c in components) {
-          final List type = c['types'];
-          if (type.contains('street_number')) {
-            place.streetNumber = c['long_name'];
-          }
-          if (type.contains('route')) {
-            place.street = c['long_name'];
-          }
-          if (type.contains('locality')) {
-            place.city = c['long_name'];
-          }
-          if (type.contains('postal_code')) {
-            place.zipCode = c['long_name'];
-          }
-        }
-        return place;
-      }
-      throw Exception(result['error_message']);
-    } else {
-      throw Exception('Failed to fetch suggestion');
-    }
   }
 }
+
