+++ libs/redux_comp/lib/actions/register_user_action.dart
@@ -2,7 +2,6 @@ import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:async_redux/async_redux.dart';
 import 'package:flutter/foundation.dart';
-import 'package:redux_comp/actions/get_address_action.dart';
 import '../app_state.dart';
 
 class RegisterUserAction extends ReduxAction<AppState> {
