+++ libs/redux_comp/lib/models/geolocation/place_model.dart
@@ -1,0 +1,18 @@
+class Place {
+  String? streetNumber;
+  String? street;
+  String? city;
+  String? zipCode;
+
+  Place({
+    this.streetNumber,
+    this.street,
+    this.city,
+    this.zipCode,
+  });
+
+  @override
+  String toString() {
+    return 'Place(streetNumber: $streetNumber, street: $street, city: $city, zipCode: $zipCode)';
+  }
+}
