+++ libs/redux_comp/lib/models/geolocation/suggestion_model.dart
@@ -1,0 +1,11 @@
+class Suggestion {
+  final String placeId;
+  final String description;
+
+  Suggestion(this.placeId, this.description);
+
+  @override
+  String toString() {
+    return 'Suggestion(description: $description, placeId: $placeId)';
+  }
+}
