+++ workspace.json
@@ -8,6 +8,7 @@
     "amplify": "libs/amplify",
     "tradesman": "libs/tradesman",
     "consumer": "libs/consumer",
-    "authentication": "libs/authentication"
+    "authentication": "libs/authentication",
+    "geolocation": "libs/geolocation"
   }
 }
